+++ copasi/compareExpressions/CNormalSum.h
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/compareExpressions/CNormalSum.h,v $
-//   $Revision: 1.10.4.1 $
+//   $Revision: 1.10.4.2 $
 //   $Name:  $
-//   $Author: shoops $
-//   $Date: 2011/01/10 17:00:27 $
+//   $Author: gauges $
+//   $Date: 2011/02/16 15:47:38 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -84,7 +84,7 @@ public:
    * Retrieve the number of summands of this sum.
    * @return int
    */
-  size_t getSize() const;
+  int getSize() const;
 
   /**
    * Add product to this sum.
